packages update to artifact; embedding service add; pom udpate
