+++ README.md
@@ -4,4 +4,4 @@ Groovy libraries for using OpenAI's API's.
 ## OpenAI APIs
 - [Models](https://platform.openai.com/docs/api-reference/models) - ModelService class
 - [Chat Completions](https://platform.openai.com/docs/api-reference/chat/create) - ChatService class
-
+- [Embedding](https://platform.openai.com/docs/api-reference/embeddings/create) - EmbeddingService class
