+++ pom.xml
@@ -4,9 +4,35 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>OpenAIGroovy</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <groupId>com.github.uszebr</groupId>
+    <artifactId>openaigroovy</artifactId>
+    <version>0.2-SNAPSHOT</version>
+
+    <name>OpenAI Groovy Library</name>
+    <description>Service classes to request OpenAI API endpoints</description>
+    <url>https://github.com/uszebr/openaigroovy</url>
+
+    <licenses>
+        <license>
+            <name>MIT License</name>
+            <url>http://www.opensource.org/licenses/mit-license.php</url>
+        </license>
+    </licenses>
+
+    <developers>
+        <developer>
+            <name>Viktor Diachuk</name>
+            <email>openaigroovy@gmail.com</email>
+            <organization>Uszebr</organization>
+            <organizationUrl>https://github.com/uszebr/</organizationUrl>
+        </developer>
+    </developers>
+
+    <scm>
+        <connection>scm:git:git://github.com/uszebr/openaigroovy.git</connection>
+        <developerConnection>scm:git:ssh://github.com:uszebr/openaigroovy.git</developerConnection>
+        <url>http://github.com/uszebr/openaigroovy</url>
+    </scm>
 
     <dependencies>
         <dependency>
