+++ src/main/groovy/com/github/uszebr/openaigroovy/chat/Role.groovy
@@ -1,0 +1,24 @@
+package com.github.uszebr.openaigroovy.chat
+
+enum Role {
+    SYSTEM('system'),
+    USER('user'),
+    ASSISTANT('assistant'),
+    FUNCTION('function')
+
+    String name
+
+    Role(String name) {
+        this.name = name
+    }
+
+    static Role readRoleFromName(String roleName) {
+        for (Role role : values()) {
+            if (role.name == roleName) {
+                return role
+            }
+        }
+        throw new IllegalArgumentException("Invalid role name: $roleName")
+    }
+
+}
