+++ src/main/groovy/com/github/uszebr/openaigroovy/model/apiResponseExample.txt
