+++ src/main/groovy/com/github/uszebr/openaigroovy/openai/OpenAIUsage.groovy
@@ -1,0 +1,13 @@
+package com.github.uszebr.openaigroovy.openai
+
+class OpenAIUsage {
+    Integer promptTokens
+    Integer completionTokens
+    Integer totalTokens
+
+    OpenAIUsage(Integer promptTokens, Integer completionTokens, Integer totalTokens) {
+        this.promptTokens = promptTokens
+        this.completionTokens = completionTokens
+        this.totalTokens = totalTokens
+    }
+}
