+++ src/main/groovy/com/github/uszebr/openaigroovy/response/FunctionCall.groovy
@@ -1,0 +1,15 @@
+package com.github.uszebr.openaigroovy.response
+
+/**
+ * Handling "function_call" for chat response message
+ */
+class FunctionCall {
+
+    String name
+    Map<String,Object> arguments
+
+    FunctionCall(String name, Map<String, Object> arguments) {
+        this.name = name
+        this.arguments = arguments
+    }
+}
