+++ src/main/groovy/com/github/uszebr/openaigroovy/response/apiResponseExampleMultiChoice.txt
