--- src/main/groovy/example/modelExample.groovy
@@ -1,41 +1,0 @@
-package example
-
-import openai.OpenAIService
-import openai.model.ModelApiResponse
-import openai.model.ModelService
-import util.ConfigReader
-
-// getting API Key from config
-ConfigReader configReader = new ConfigReader('/config.properties')
-def apiKey = configReader.readConfig().apiKey.toString()
-
-// setting up OpenAIService
-// timeouts can be set here
-OpenAIService service = OpenAIService.builder().withApiKey(apiKey).build()
-
-// creating ModalService
-ModelService modelAPI = new ModelService(service)
-// calling OpenAI model api
-ModelApiResponse modelApiResponse = modelAPI.call()
-// checking data response data
-println modelApiResponse.isResponse() // true
-// getting AI model objects from raw data
-def models = modelApiResponse.getAllModels()
-
-// printing all model names
-models.each { println(it.id) }
-//text-search-babbage-doc-001
-//gpt-3.5-turbo-16k-0613
-//curie-search-query
-//gpt-3.5-turbo
-//gpt-3.5-turbo-16k
-//text-search-babbage-query-001
-//babbage
-//babbage-search-query
-//text-babbage-001
-//whisper-1
-//text-similarity-davinci-001
-//gpt-4
-//...
-
-
