--- src/main/groovy/openai/chat/response/FunctionCall.groovy
@@ -1,15 +1,0 @@
-package openai.chat.response
-
-/**
- * Handling "function_call" for chat response message
- */
-class FunctionCall {
-
-    String name
-    Map<String,Object> arguments
-
-    FunctionCall(String name, Map<String, Object> arguments) {
-        this.name = name
-        this.arguments = arguments
-    }
-}
